[tiny] split middleware hook to separate file
